Integrated external needs test case
